remove shim layers
