+++ spark2-sql-plugin/src/main/scala/org/apache/spark/sql/rapids/collectionOperations.scala
